refactor: product package 분리
