+++ src/main/java/kitchenpos/menu/Menu.java
@@ -7,6 +7,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import kitchenpos.common.Price;
 
 @Entity
 public class Menu {
