+++ src/main/java/kitchenpos/menu/MenuProducts.java
@@ -8,6 +8,7 @@ import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.FetchType;
 import javax.persistence.OneToMany;
+import kitchenpos.common.Price;
 
 @Embeddable
 public class MenuProducts {
