+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.menu.repository;
+package kitchenpos.product.repository;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.menu.Product;
+import kitchenpos.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
