+++ src/test/java/kitchenpos/acceptance/MenuAcceptanceTest.java
@@ -5,8 +5,8 @@ import java.util.Collections;
 import java.util.List;
 import kitchenpos.menu.MenuGroup;
 import kitchenpos.menu.MenuProduct;
-import kitchenpos.menu.Product;
 import kitchenpos.menu.dto.MenuRequest;
+import kitchenpos.product.Product;
 import kitchenpos.support.RequestBuilder;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.DisplayName;
