+++ src/test/java/kitchenpos/acceptance/OrderAcceptanceTest.java
@@ -3,7 +3,6 @@ package kitchenpos.acceptance;
 import io.restassured.response.ValidatableResponse;
 import kitchenpos.menu.MenuGroup;
 import kitchenpos.menu.MenuProduct;
-import kitchenpos.menu.Product;
 import kitchenpos.order.MenuInfo;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderLineItem;
@@ -11,6 +10,7 @@ import kitchenpos.order.OrderStatus;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderStatusRequest;
 import kitchenpos.ordertable.OrderTable;
+import kitchenpos.product.Product;
 import kitchenpos.support.RequestBuilder;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.DisplayName;
