+++ src/test/java/kitchenpos/common/PriceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu;
+package kitchenpos.common;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
