+++ src/test/java/kitchenpos/product/ProductTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.menu;
+package kitchenpos.product;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
+import kitchenpos.common.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
